Improve stream notification handling
This makes it cache the stream status to properly work with both twitch and Discord's inbuilt streaming in unison. It also adds back the ability to specify the game name (by querying Twitch's API for the game info) and fixes an issue where the stream marker wasn't removed from a voice channel.